refactor(types): Centralize types in `types.ts`
