+++ package/src/index.ts
@@ -1,16 +1,13 @@
-export { ShikiHighlighter as default } from './component';
 export { useShikiHighlighter } from './hook';
-export type { ShikiHighlighterProps } from './component';
+export { ShikiHighlighter as default } from './component';
 
-export {
-  isInlineCode,
-  rehypeInlineCodeProperty,
-  type Element,
-} from './utils';
+export { isInlineCode, rehypeInlineCodeProperty } from './utils';
 
 export type {
   Language,
   Theme,
   Themes,
+  Element,
   HighlighterOptions,
+  ShikiHighlighterProps,
 } from './types';
