+++ package/src/utils.ts
@@ -1,14 +1,15 @@
 import { visit } from 'unist-util-visit';
 import { bundledLanguages, isSpecialLang } from 'shiki';
-import type { ShikiTransformer, ThemeRegistrationAny } from 'shiki';
-import type { Element } from 'hast';
-import type { Language, Theme, Themes, TimeoutState } from './types';
-import type { LanguageRegistration } from './extended-types';
 
-/**
- * Type for the HAST node, use to type `node` from react-markdown
- */
-export type { Element };
+import type { ShikiTransformer, ThemeRegistrationAny } from 'shiki';
+import type { LanguageRegistration } from './extended-types';
+import type {
+  Language,
+  Theme,
+  Themes,
+  TimeoutState,
+  Element,
+} from './types';
 
 /**
  * Rehype plugin to add an 'inline' property to <code> elements
